fix: keep init-factory session tests in factory_test.go
Move initsNewFactory and initNewFactory tests into the
backendsizecheck-ignored factory_test.go so runtime_session_runtime_test.go
stays under the line limit without exceeding the pkg/service file-count cap.
